Small package refactor (preparation for proguard)
