+++ app/src/main/java/pl/droidcon/app/service/api/AgendaService.java
@@ -1,6 +1,6 @@
 package pl.droidcon.app.service.api;
 
-import pl.droidcon.app.service.model.AgendaResponse;
+import pl.droidcon.app.model.api.AgendaResponse;
 import retrofit.http.GET;
 
 public interface AgendaService {
